Update : Removed AWSSDK and PHP + Testing Database functionality
Levels can be queried by name and return a json object which is then saved into the user's persistent data files.

This is meant to serve as a POC for combined C#/PHP usage.

Will need to seek out a manner in which to deploy PHP code onto a server. Check other AWS free-tier services (?)