Head stuck in the game. Gotta pull back and look at the bigger picture.
Adding tons of poorly thought-out accessors and methods to classes. It's becoming impossible to remember which component knows what. Or how to properly instantiate and place cubes. Or how to move them, when the time comes. Or reset them. Or rewind. The list goes on, it would do me good to sit back and review exactly what is going on here and go back to the drawing board instead of progressing in the dark any further.
I'm in a good place right now. There's a working Level Editor, with saving and loading implemented. The next steps are to get all of these accessors under control, determine who is doing what, and progress into loading levels in game. Once that's done, it's on to basic gameplay programming. At that point, the CubeBehaviorManagers are going to come into play, which is bound to be a blast.

Anyway, checking out now. I'll see you in a minute.