Game Camera is another giant problem. I don't know how I'll finish this.
Seriously, I feel like every step I take, I'm thrown back 12. I just spent the entire day and all of last night trying to get movement up and running, and I'm still not satisfied with the result. The code isn't clean - hardly legible, it's full of small necessary patchwork fixes that are necessary to get the machine running. All of that frustration later, only to discover that maybe non of what I'm doing is really that efficient. The Game Camera is going to be a massive pain because I've locked myself in this tiny shitty little box with constraints left and right. The simple solution would be to rotate the Kubo struture on its Y axis, but I can't do that because then none of the cubes know where the fuck they are anymore, and things go haywire.
The solution seems not only complex, but way out of reach for the moment. All this for a mechanic that should be simple. Trivial. Nothing to even really worry about.

It's killing me bruh.